Вместо заглушки вывода статистики, теперь выводится информация о кол-ве животных и растений

+++ island/src/main/java/com/javarush/island/ostapenko/core/dto/SimulationStatistics.java
@@ -1,25 +1,25 @@
 package com.javarush.island.ostapenko.core.dto;
 
 public class SimulationStatistics {
-    private final int currentDay;
-    private final int totalAnimal;
-    private final int totalPlants;
+    private final long currentDay;
+    private final long totalAnimal;
+    private final long totalPlants;
 
-    public SimulationStatistics(int currentDay, int totalAnimal, int totalPlants) {
+    public SimulationStatistics(long currentDay, long totalAnimal, long totalPlants) {
         this.currentDay = currentDay;
         this.totalAnimal = totalAnimal;
         this.totalPlants = totalPlants;
     }
 
-    public int getCurrentDay() {
+    public long getCurrentDay() {
         return currentDay;
     }
 
-    public int getTotalAnimal() {
+    public long getTotalAnimal() {
         return totalAnimal;
     }
 
-    public int getTotalPlants() {
+    public long getTotalPlants() {
         return totalPlants;
     }
 }
